requirements.txt
click
rich
RPi.GPIO
pyyaml
